msm8660: update libril
* Sync with https://github.com/CyanogenMod/android_hardware_samsung/commits/b8400a23f559563a5b652dfc88bd951cd6a2e7ae/ril/libril

libril: Fix network operator search

Change-Id: Ib904875c403942ec96dfa229080d9543d7c9b8fa

libril: Fix element count

This lines should drop every 5th element (#5,#10,#15,...).
At the moment it is dropping elements #6, #11, #16, as it start from 0
and the list start from 1. So let's match it up.

Thanks to Tobias Heinemann for the idea.

Change-Id: Idc2904040119716e666ea0fe320cde1d45869ee5

ril: Fix compilation with debug messages turned on

Change-Id: Ifabf416ebdfe681e63df1d127567d0ef54cfc0ec
Signed-off-by: Andreas Schneider <asn@cryptomilk.org>

ril: Support new sound manger RIL responses

New version of Samsung RIL do not use RIL_UNSOL_WB_AMR_STATE anymore.
There is a Sound Manager (SND-MGR) which has its own response base
(20000). So to be able to register a handler for

RIL_UNSOL_SNDMGR_WB_AMR_REPORT

with libsecril-client we need to know about it. This allows us to
register a callback in the audio HAL to turn WB support on.

WB_AMR_REPORT report 0x00, 0x01 or 0x02. I guess 0x01 is WB ON and
0x02 is WB + NS ON.

Change-Id: I403c86781551fc3b08354187322e8443402d56f6

Removal of dead code and adding log messages to make it easier
for debugging

- Add more logging using ALOGD(..) for function dispatchSmsWrite,
dispatchDial, dispatchSIM_IO, dispatchSIM_APDU, dispatchCallForward,
dispatchRaw and dispatchCdmaSmsAck. To make it easier for debugging.

- Remove commented log message inside wakeTimeoutCallback () function
and remove the else clause as it's dead not used for anything.

- Modify log message inside removeFromList and removeWatch using +
and - to make it consistent with the other log message.

Change-Id: Ife0212e6a4724d8b06a1dd767484bfdf98b8e1a6
Signed-off-by: Nanik Tolaram <nanikjava@gmail.com>

Introduction of the new SAP-UIM socket

Added new socket connection for communication between the SAP module
in Telephony and User Identity Module with DSDA support.

New classes added are:

RilSocket
---------
Base class(virtual) for any type of socket. Has listen
and commands callbacks similar to that of the existing
telephony->rild socket.

RilSapSocket
------------
Derived class for socket for communication between BT SAP
and the sim module in the modem. It initialises
socket, calls Sap_Uim_Init to get the handles and
starts the socket thread.

rilSocketQueue
--------------
Queue for managing socket requests.

Change-Id: I8828173941d6ae76f1f9cc0d567efaf41a77d175

libril: tag RILC messages to radio log

Change-Id: Ia3234aa933249eb9f383c7e45c3f67b3c0c50634

Remove verbose logs

Several things are redundent except for the fact
messages are reaching rilc.  Left in in case we need to debug
the layer, but blocked out with #if VDBG

Reduces startup logging to radio buffer by 20%

Change-Id: I95099457ea365ee108e7da8e1f81434531b6b9f0

ril: Fix network operator search

Newer RIL versions have more QAN elements so we need a way
to adjust them to the correct value.

This needs to be set to same value as mQANElements in the RIL Telephony
class.

Change-Id: Icfad9c0acf1b95f63c232b6c4c49ed52bd4e3aa4
Signed-off-by: Andreas Schneider <asn@cryptomilk.org>

ril: Use debug log level and not error

Change-Id: I8f017cb181967790260183fd06d1226d303bd2b8
Signed-off-by: Andreas Schneider <asn@cryptomilk.org>

libril: fix compilation with debugging enabled

Change-Id: Iac159175a23b2810e56c38fb82cae58c20dddb8e

ril: Unify network elements property check

frameworks/opt/telephony is using a slightly different named property
to check for the QANElements count.
Align the check here to the frameworks one to avoid redundancies.

Change-Id: I22f7cf070953026798d0ae4759e3c9845dba767c